% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw2011-data.R
\docType{data}
\name{kw2011}
\alias{kw2011}
\title{municipal waste in Kanto region, Japan in 2011
 
 A dataset about municipal waste in Kanto region of Japan,
 which includes total amount of waste, the breakdown of them,
 such as waste from household and buisiness activities,
 and recycle ratio.
 
 \itemize{
   \item pref. name of the prefecture
   \item citycode. code of the municipality
   \item population. population of the municipality
   \item total.W. total amount of waste
   \item life.W. amount of waste from household
   \item industrial.W. amount of waste from business activities
   \item RR. recycle ratio
   \item cityname. cityname of the municipality
 }}
\format{A data frame with 287 rows and 8 variables}
\usage{
data(kw2011)
}
\description{
municipal waste in Kanto region, Japan in 2011
 
 A dataset about municipal waste in Kanto region of Japan,
 which includes total amount of waste, the breakdown of them,
 such as waste from household and buisiness activities,
 and recycle ratio.
 
 \itemize{
   \item pref. name of the prefecture
   \item citycode. code of the municipality
   \item population. population of the municipality
   \item total.W. total amount of waste
   \item life.W. amount of waste from household
   \item industrial.W. amount of waste from business activities
   \item RR. recycle ratio
   \item cityname. cityname of the municipality
 }
}
\references{
\url{http://www.env.go.jp/recycle/waste_tech/}
}
\keyword{datasets}
