% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdmR-package.R
\docType{package}
\name{vdmR-package}
\alias{vdmR-package}
\alias{vdmR}
\title{vdmR: Visual Data Mining Tools for R}
\description{
Visual Data Mining Tools for R
}
\details{
This package provides web based visual data mining tools by adding
interactive functions to ggplot2 graphics. Brushing and linking between the
multiple plots is one of the main feature of this package. Currently
scatter plot, histogram, parallel coordinate plot and choropleth map are
supported in this package.
}
\references{
\insertRef{Fujino2017}{vdmR}
}
\author{
Author: Tomokazu Fujino.
Maintainer: Tomokazu Fujino. \email{fujino@fwu.ac.jp}
}
