% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlaunch.R
\docType{methods}
\name{vlaunch}
\alias{vlaunch}
\title{Generate main window for interactive plot windows}
\usage{
vlaunch(data, name, tag, path = tempdir(), iframe = FALSE, browse = TRUE)
}
\arguments{
\item{data}{data frame for default data set}

\item{name}{character for the name of the generated scatter plot}

\item{tag}{character for the common name of a series of linked plots}

\item{path}{character string of a directory for writing HTML and SVG files}

\item{iframe}{logical; if \code{TRUE}, all plot windows are displayed in the main window as inline frames}

\item{browse}{logical; if \code{TRUE}, browse the main window by the default web browser through the local web server; if \code{FALSE}, generating only}
}
\description{
\code{vlauch} generates a main window which opens each pre-generated window including statistical plot with interactivity
}
\examples{
data(vsfuk2012)
vscat(MarriageRate, DivorceRate, vsfuk2012, "scat1", "vsfuk2012", colour=Type)
vhist(FertilityRate, vsfuk2012, "hist1", "vsfuk2012", fill=Type)
vlaunch(vsfuk2012, "main", "vsfuk2012", browse=FALSE)

}
