% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vscat.R
\docType{methods}
\name{vscat}
\alias{vscat}
\title{Generate scatter plot with interactive functions}
\usage{
vscat(x, y, data, name, tag, path = tempdir(), ...)
}
\arguments{
\item{x, y}{column name of data frame \code{data} for x-axis and y-axis of scatter plot}

\item{data}{data frame for default data set}

\item{name}{character for the name of the generated scatter plot}

\item{tag}{character for the common name of a series of linked plots}

\item{path}{character string of a directory for writing HTML and SVG files}

\item{...}{aesthetic mappings to be passed to ggplot2 methods}
}
\description{
\code{vscat} generates scatter plot of variable \code{x} and \code{y} of data frame \code{data} with interactive functions.
}
\examples{
data(vsfuk2012)
vscat(MarriageRate, DivorceRate, vsfuk2012, "scat1", "vsfuk2012", colour=Type)
vlaunch(vsfuk2012, "main", "vsfuk2012", browse=FALSE)

}
